unite id type to Integer from Long

+++ backend/src/main/java/ttajiri/todayeats/repository/CategoryRepository.java
@@ -3,5 +3,5 @@ package ttajiri.todayeats.repository;
 import org.springframework.data.repository.*;
 import ttajiri.todayeats.repository.dto.*;
 
-public interface CategoryRepository extends CrudRepository<CategoryDto, Long> {
+public interface CategoryRepository extends CrudRepository<CategoryDto, Integer> {
 }
